Add windows test runner
os.Getwd() in utils.go behaves differently in windows to mac/linux and had to be swapped out for runtime.Caller.  Also exec.Command refused to run any file without a .exe file extension on windows.  Wasmer has been omitted from windows builds, as it doesn't compile - I was wrong in the standup, the dependency in go.mod is ignored if it is never used, this was hidden behind the two other windows issues, and the fact that the windows build logs out more than 110,000 lines of compiler warnings (I think this is unavoidable, the wasmtime ci does the same as far as I can see).