ci: Add tests workflow (#63)
* Make validate title workflow run against main

develop doesn't exist in this repo...

* Add correct license header to title workflow

* Re-organise make files

Splits out deps, and adds test:ci using gotestsum. Also restructures the test calls so that all deps are built once at the start and then the unit and integration tests get executed one after the other at the end - logs/output is much nicer this way.

* Add test workflow

* Add windows test runner

os.Getwd() in utils.go behaves differently in windows to mac/linux and had to be swapped out for runtime.Caller.  Also exec.Command refused to run any file without a .exe file extension on windows.  Wasmer has been omitted from windows builds, as it doesn't compile - I was wrong in the standup, the dependency in go.mod is ignored if it is never used, this was hidden behind the two other windows issues, and the fact that the windows build logs out more than 110,000 lines of compiler warnings (I think this is unavoidable, the wasmtime ci does the same as far as I can see).