Extract native module from test
This will be used shortly by more tests.  The extraction is still really really minimal - it does what it needs to do now and is no fancier.  In the future we may want to do more with it, but not now.  The code is a cut and paste, plus fixing the references to memory.