Clone input_vec before parsing
And clone the output.  The changes to type_rdr and len_rdr are not technically required, but they are more descriptive and consistent with the input_vec.

The Clone generic constraint is a new restriction - however that should be implemented by basically anything anyone would want to transform using lens (famous last words...).  It could be handled outside of this function (it is required for `StreamOption::Some(v.clone())`), however if that is not handled here, the lens author must do so lest they risk memory corruption problems - I feel like the clone constraint is a much lesser evil.