Mark try_from_mem as unsafe
Because it is, and should have been makred as such when I first introduced it. The unsafes within this function *could* be removed, but I kind of like them as they flag the particularly unsafe parts of the function, they are also compiler warnings in later editions (not the currently used 2021).